Refactor app.ts into smaller, modular components
Break up the monolithic app.ts file into smaller, more focused modules:
- compilation-env.ts: Initialization of compilation environment
- compiler-changes.ts: Compiler change event handling
- compiler-discovery.ts: Compiler discovery and loading
- controllers.ts: API controller setup
- rendering.ts: Template rendering
- routes-setup.ts: Express routes and API setup
- server-config.ts: Server configuration middleware
- server-listening.ts: Server listening setup
- static-assets.ts: Static file handling
- temp-dir.ts: Temporary directory setup
- url-handlers.ts: URL handling and redirects

Each module has a single responsibility, making the code more maintainable
and testable. Added appropriate tests for the new modules.
