+++ lib/app/compilation-env.ts
@@ -1,0 +1,75 @@
+// Copyright (c) 2025, Compiler Explorer Authors
+// All rights reserved.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+//
+//     * Redistributions of source code must retain the above copyright notice,
+//       this list of conditions and the following disclaimer.
+//     * Redistributions in binary form must reproduce the above copyright
+//       notice, this list of conditions and the following disclaimer in the
+//       documentation and/or other materials provided with the distribution.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
+
+import type {AppArguments} from '../app.interfaces.js';
+import {CompilationEnvironment} from '../compilation-env.js';
+import {CompilationQueue} from '../compilation-queue.js';
+import {FormattingService} from '../formatting-service.js';
+import {CompileHandler} from '../handlers/compile.js';
+import {PropertyGetter} from '../properties.interfaces.js';
+import {CompilerProps} from '../properties.js';
+
+export interface CompilationEnvResult {
+    compilationQueue: CompilationQueue;
+    compilationEnvironment: CompilationEnvironment;
+    compileHandler: CompileHandler;
+    formattingService: FormattingService;
+}
+
+/**
+ * Initialize the compilation environment components
+ * @param appArgs - Application arguments
+ * @param compilerProps - Compiler properties
+ * @param ceProps - Compiler Explorer properties
+ * @param awsProps - AWS properties
+ * @returns Object containing initialized compilation components
+ */
+export async function initializeCompilationEnvironment(
+    appArgs: AppArguments,
+    compilerProps: CompilerProps,
+    ceProps: PropertyGetter,
+    awsProps: PropertyGetter,
+): Promise<CompilationEnvResult> {
+    const formattingService = new FormattingService();
+    await formattingService.initialize(ceProps);
+
+    const compilationQueue = CompilationQueue.fromProps(compilerProps.ceProps);
+    const compilationEnvironment = new CompilationEnvironment(
+        compilerProps,
+        awsProps,
+        compilationQueue,
+        formattingService,
+        appArgs.doCache,
+    );
+
+    const compileHandler = new CompileHandler(compilationEnvironment, awsProps);
+    compilationEnvironment.setCompilerFinder(compileHandler.findCompiler.bind(compileHandler));
+
+    return {
+        compilationQueue,
+        compilationEnvironment,
+        compileHandler,
+        formattingService,
+    };
+}
